+++ src/Git_Updater/API/GitHub_API.php
@@ -286,6 +286,7 @@ class GitHub_API extends API implements API_Interface {
 			function ( $e ) use ( &$arr ) {
 				$arr['private']      = $e->private;
 				$arr['last_updated'] = $e->pushed_at;
+				$arr['added']        = $e->created_at;
 				$arr['watchers']     = $e->watchers;
 				$arr['forks']        = $e->forks;
 				$arr['open_issues']  = $e->open_issues;