+++ src/Git_Updater/Init.php
@@ -74,13 +74,6 @@ class Init {
 		add_action( 'init', [ $this->base, 'background_update' ] );
 		add_action( 'init', [ $this->base, 'set_options_filter' ] );
 
-		add_action( 'deprecated_hook_run', [ new Messages(), 'deprecated_error_message' ], 10, 4 );
-
-		// `wp_get_environment_type()` added in WordPress 5.5.
-		if ( function_exists( 'wp_get_environment_type' ) && 'development' === wp_get_environment_type() ) {
-			add_filter( 'deprecated_hook_trigger_error', '__return_false' );
-		}
-
 		// Load hook for adding authentication headers for download packages.
 		add_filter(
 			'upgrader_pre_download',