+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update to correct format of readme tags
+* add correctly formated date/time for `update-api` REST endpoint
 
 #### 12.14.0 / 2025-02-26
 * make sure proper release asset headers are added even if access token not set