+++ CHANGES.md
@@ -2,6 +2,7 @@
 * update to correct format of readme tags
 * add correctly formated date/time for `update-api` REST endpoint
 * remove deprecated hooks from v10 and earlier
+* add error checking to `parse_contents_response()`
 
 #### 12.14.0 / 2025-02-26
 * make sure proper release asset headers are added even if access token not set