+++ CHANGES.md
@@ -3,6 +3,7 @@
 * add correctly formated date/time for `update-api` REST endpoint
 * remove deprecated hooks from v10 and earlier
 * add error checking to `parse_contents_response()`
+* update Freemius/wordpress-sdk
 
 #### 12.14.0 / 2025-02-26
 * make sure proper release asset headers are added even if access token not set